Migrate '/schema/zkversion' to api module
Also merges GetSchemaAPI and GetSchemaZkVersionAPI, since there's no
real need for them to be separate.
