+++ solr/api/src/java/org/apache/solr/client/api/model/SchemaInfoResponse.java
@@ -17,14 +17,13 @@
 package org.apache.solr.client.api.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.Map;
 
 public class SchemaInfoResponse extends SolrJerseyResponse {
-    // TODO The schema response is quite complicated, so for the moment it's sufficient to record it
-    // here only as a Map.  However, if SOLR-16825 is tackled then there will be a lot of value in
-    // describing this response format more accurately so that clients can navigate the contents
-    // without lots of map fetching and casting.
-    @JsonProperty("schema")
-    public Map<String, Object> schema;
+  // TODO The schema response is quite complicated, so for the moment it's sufficient to record it
+  // here only as a Map.  However, if SOLR-16825 is tackled then there will be a lot of value in
+  // describing this response format more accurately so that clients can navigate the contents
+  // without lots of map fetching and casting.
+  @JsonProperty("schema")
+  public Map<String, Object> schema;
 }
