+++ solr/api/src/java/org/apache/solr/client/api/util/Constants.java
@@ -24,7 +24,8 @@ public class Constants {
 
   public static final String INDEX_TYPE_PATH_PARAMETER = "indexType";
   public static final String INDEX_NAME_PATH_PARAMETER = "indexName";
-  public static final String INDEX_PATH_PREFIX = "/{" + INDEX_TYPE_PATH_PARAMETER + ":cores|collections}/{" + INDEX_NAME_PATH_PARAMETER + "}";
+  public static final String INDEX_PATH_PREFIX =
+      "/{" + INDEX_TYPE_PATH_PARAMETER + ":cores|collections}/{" + INDEX_NAME_PATH_PARAMETER + "}";
 
   public static final String BINARY_CONTENT_TYPE_V2 = "application/vnd.apache.solr.javabin";
 }
