+++ solr/api/src/java/org/apache/solr/client/api/util/IndexApiParameters.java
@@ -16,27 +16,32 @@
  */
 package org.apache.solr.client.api.util;
 
+import static org.apache.solr.client.api.util.Constants.INDEX_NAME_PATH_PARAMETER;
+import static org.apache.solr.client.api.util.Constants.INDEX_TYPE_PATH_PARAMETER;
+
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Schema;
-import org.apache.solr.client.api.model.IndexType;
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-
-import static org.apache.solr.client.api.util.Constants.INDEX_NAME_PATH_PARAMETER;
-import static org.apache.solr.client.api.util.Constants.INDEX_TYPE_PATH_PARAMETER;
+import org.apache.solr.client.api.model.IndexType;
 
 /**
- * Concisely collects the parameters shared by APIs that interact with contents of a specific core/collection.
+ * Concisely collects the parameters shared by APIs that interact with contents of a specific
+ * core/collection.
  *
- * Used primarily as a way to avoid duplicating these parameter definitions on each relevant interface method in {@link org.apache.solr.client.api.endpoint}
+ * <p>Used primarily as a way to avoid duplicating these parameter definitions on each relevant
+ * interface method in {@link org.apache.solr.client.api.endpoint}
  */
-// TODO NOCOMMIT Reconsider naming here - do we have a generic term to refer to both collections and cores.  "Index" is waaaay too overloaded.
+// TODO NOCOMMIT Reconsider naming here - do we have a generic term to refer to both collections and
+// cores.  "Index" is waaaay too overloaded.
 @Target({ElementType.METHOD, ElementType.TYPE, ElementType.PARAMETER, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-@Parameter(name = INDEX_TYPE_PATH_PARAMETER, in = ParameterIn.PATH, schema = @Schema(implementation = IndexType.class, enumAsRef = true))
+@Parameter(
+    name = INDEX_TYPE_PATH_PARAMETER,
+    in = ParameterIn.PATH,
+    schema = @Schema(implementation = IndexType.class, enumAsRef = true))
 @Parameter(name = INDEX_NAME_PATH_PARAMETER, in = ParameterIn.PATH)
 public @interface IndexApiParameters {}
